pyproject.toml
[tool.poetry]
name = "duckdb-version-manager"
version = "0.1.0"
description = ""
authors = ["Niclas Haderer <niclas.haderer.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pygithub = "^2.2.0"
black = "^24.2.0"
httpx = "^0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

